Send Content-Length when sending a file.
Send the Content-Length header when sending a file so that when a large file is
sent the browser can inform the user how long it will take for the file to be
fully downloaded.

TODO - If we really want the webserver to be able to send a large file, we want
it to be piped.  At the moment, the large file is fully read into the memory
then sent to the user which is not ideal.