Use less memory.
Previously, when sending a file the entirety of the file was read into memory
then sent to the recipient.  This could use up a lot of memory when sending a
large file.

Now the file is sent in smaller chunks, so even sending a large file does not
take up too much memory.

TODO - It would be nice to handle new connections between sending the chunks.
At the moment no new connections are handled while a large file is being sent.